refactored the files and added PostsPage.js
