+++ instagram-app/src/App.js
@@ -1,32 +1,13 @@
-import dummyData from './dummy-data';
 import React from 'react';
 import './App.scss';
-import SearchBar from './components/SearchBar/SearchBar';
-import PostContainer from './components/PostContainer/PostContainer';
+import PostsPage from './components/PostContainer/PostsPage';
 
 class App extends React.Component {
-
-  state = {
-    postData: [],
-    filteredPosts: []
-  };
-
-  componentDidMount() {
-    this.setState({
-      postData: dummyData,
-    });
-  }
-
-  searchFilter = e => {
-    const filtered = this.state.postData.filter(post => post.username.toLowerCase().includes(e.target.value.toLowerCase()))
-    this.setState({ filteredPosts: filtered })
-  }
 
   render() {
     return (
       <div className="App">
-        <SearchBar searchFilter={this.searchFilter} />
-        <PostContainer filteredPosts={this.state.filteredPosts} postData={this.state.postData} />
+        <PostsPage />
       </div>
     );
   }
